move : display transaction to components
